Gui changed and boundary values changed
Now user can see their hands from the two cameras of the Leap Motion
device

The interface also allows the user to see the raw data x, y, and z from
the leap motion and the data sent to the pincher robot for comparison.

The boundaries were also reset such that it was more conservative so
the pincher robot did not revert to the default position if it was sent
a value out of range for the robot. This increases the interface
between the user and the robot as it makes the movement more smooth.